+++ backend/src/main/java/com/bookbook/domain/review/repository/ReviewRepository.java
@@ -9,9 +9,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ReviewRepository extends JpaRepository<Review, Integer> {
+public interface ReviewRepository extends JpaRepository<Review, Long> {
     
-    Optional<Review> findByRentIdAndReviewerId(Integer rentId, Long reviewerId);
+    Optional<Review> findByRentIdAndReviewerId(Long rentId, Long reviewerId);
 
     @Query("SELECT AVG(r.rating) FROM Review r WHERE r.revieweeId = :revieweeId")
     Optional<Double> findAverageRatingByRevieweeId(@Param("revieweeId") Long revieweeId);
